+++ bluepyemodel/evaluation/evaluator.py
@@ -464,9 +464,9 @@ def define_threshold_based_optimisation_protocol(
         strict_holding_bounds (bool): to adaptively enlarge bounds if holding current is outside
             when set to False
         use_fixed_dt_recordings (bool): whether to record at a fixed dt of 0.1 ms.
-        max_depth_holding_search (float): maximum depth for the binary search for the
+        max_depth_holding_search (int): maximum depth for the binary search for the
             holding current
-        max_depth_threshold_search (float): maximum depth for the binary search for the
+        max_depth_threshold_search (int): maximum depth for the binary search for the
             threshold current
     """
